% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmas3-board-class.R
\name{add_move}
\alias{add_move}
\alias{add_move.xmas3plot}
\title{Add Move to Board Plot}
\usage{
add_move(x, ...)

\method{add_move}{xmas3plot}(p, move, arr = arrow(), ...)
}
\arguments{
\item{...}{passed along (for color, linetype etc.)}

\item{p}{Board Plot}

\item{move}{Move to Add}

\item{arr}{arrow to use}
}
\value{
plot with move added
}
\description{
Add Move to Board Plot
}
