% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmas3-board-class.R
\name{add_tile.xmas3plot}
\alias{add_tile.xmas3plot}
\title{Add Tile to Board Plot}
\usage{
\method{add_tile}{xmas3plot}(p, pos, ...)
}
\arguments{
\item{p}{}

\item{...}{passed on}

\item{col}{A-J or 1:10}

\item{row}{row in question}
}
\description{
Add Tile to Board Plot
}
