% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent-0.R
\name{agent_0}
\alias{agent_0}
\title{Agent 0}
\usage{
agent_0(B, n = 100, ...)
}
\arguments{
\item{B}{Board Matrix}

\item{n}{number of moves to make}

\item{...}{not in use for this agent}
}
\value{
a data.frame containing as least move_sequence
}
\description{
Makes random moves
}
