% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent-4.R
\name{agent_4}
\alias{agent_4}
\title{Agent 4}
\usage{
agent_4(B, ...)
}
\arguments{
\item{B}{Board Matrix}

\item{...}{not in use for this agent}
}
\value{
a data.frame containing as least move_sequence
}
\description{
Plays a 4x4 Best Move Strategy
}
