% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent-5.R
\name{agent_5}
\alias{agent_5}
\title{Agent 5}
\usage{
agent_5(B, ...)
}
\arguments{
\item{B}{Board Matrix}

\item{...}{not in use for this agent}
}
\value{
a data.frame containing as least move_sequence
}
\description{
Plays a 5x3 Quarter
}
