% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent-6.R
\name{agent_6}
\alias{agent_6}
\title{Agent 6}
\usage{
agent_6(B, ...)
}
\arguments{
\item{B}{Board Matrix}

\item{...}{not in use for this agent}
}
\value{
a data.frame containing as least move_sequence
}
\description{
Plays a 5x3 grid strategy
}
