% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent-7.R
\name{agent_7}
\alias{agent_7}
\title{Agent 7}
\usage{
agent_7(B, ...)
}
\arguments{
\item{B}{Board Matrix}

\item{...}{not in use for this agent}
}
\value{
a data.frame containing as least move_sequence
}
\description{
Plays a 5x3 grid strategy
}
