% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array-to-matrix.R
\name{array_to_LaTeX}
\alias{array_to_LaTeX}
\title{Format Matrix or Array to Latex}
\usage{
array_to_LaTeX(arr)
}
\arguments{
\item{arr}{}
}
\value{
Latex code
}
\description{
Helper function to generate latex tables from matrices or arrays.
}
\author{
Gregory Janesch
}
