% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmas3-board-class.R
\name{as.matrix.xmas3}
\alias{as.matrix.xmas3}
\title{Converts Board State to Matrix}
\usage{
\method{as.matrix}{xmas3}(x)
}
\arguments{
\item{x}{board}
}
\value{
matrix
}
\description{
Converts Board State to Matrix
}
