% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascii-art.R
\name{ascii_gatai}
\alias{ascii_gatai}
\title{Get ASCII from the Collection}
\usage{
ascii_gatai(
  whichone = c("default", "scared", "surprised", "cool", "cash", "computer", "isee",
    "bigtree", "tree", "bigtreebroken", "bigtreedown", "treedown", "hiding")
)
}
\arguments{
\item{whichone}{}
}
\value{
character vector with the art
}
\description{
Get ASCII from the Collection
}
\examples{
ascii_gatai()
}
