% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binarize-board.R
\name{binarize_board}
\alias{binarize_board}
\title{Make Board Binary}
\usage{
binarize_board(Board, tile = LETTERS[1:10])
}
\arguments{
\item{Board}{}

\item{tile}{}
}
\value{
a character matrix with "0" and "1"
}
\description{
Turns the board matrix into a binary matrix based in the input letter
}
