% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-binary-to-matrix.R
\name{binary_to_matrix}
\alias{binary_to_matrix}
\title{Convert binary string back to matrix}
\usage{
binary_to_matrix(x)
}
\arguments{
\item{x}{}
}
\value{
matrix
}
\description{
Convert binary string back to matrix
}
