% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-centroid-coordinates.R
\name{calculate_grid_centroids}
\alias{calculate_grid_centroids}
\title{Get Centroids on Board}
\usage{
calculate_grid_centroids(
  surface_width = 1450,
  surface_height = 890,
  num_rows = 6,
  num_cols = 10
)
}
\arguments{
\item{surface_width}{width of board}

\item{surface_height}{height of board}

\item{num_rows}{number of rows}

\item{num_cols}{number of rows}
}
\value{
data.frame with col,row,x and y
}
\description{
Calculates the centroids on the Board
}
