% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-windowed-move.R
\name{calculate_windowed_move}
\alias{calculate_windowed_move}
\title{Calculate Windowed Move}
\usage{
calculate_windowed_move(
  .c,
  .r,
  move = c("First", "Last", "Middle"),
  shape = c("1x4", "2x3"),
  direction = c("Horizontal", "Vertical")
)
}
\arguments{
\item{move}{Move to make}

\item{shape}{shape of the window}

\item{direction}{vertical or horizontal}

\item{.col}{column}

\item{.row}{row}
}
\value{
coordinates for the move function and move (UDLR)
}
\description{
Convert a move from recommended for a specific window to one that is playable on the board.
}
