% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify-tile.R
\name{classify_tile}
\alias{classify_tile}
\title{Classify Tile}
\usage{
classify_tile(tile)
}
\arguments{
\item{tile}{either a character vector (filename) or magick-image}
}
\value{
a data.frame with the prediction from each model
}
\description{
Classify Tile
}
