% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-analysis.R
\name{color_analysis}
\alias{color_analysis}
\title{Analyze Color in an Image}
\usage{
color_analysis(img)
}
\arguments{
\item{img}{either a "magick-image or character-vector (filename)}
}
\value{
a data.frame with variables
}
\description{
Analyzes the colors of an image and returns overall mean along with means for the quadrants.
Reduces the color space to RBG.
}
