% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_action_space.R
\name{create_action_space}
\alias{create_action_space}
\title{Create Action Space}
\usage{
create_action_space(nrows, ncols)
}
\arguments{
\item{nrows}{board height}

\item{ncols}{board width}
}
\value{
a character vector with the possible moves
}
\description{
Creates an action space for a board (grid of nrows by ncols)
}
