% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-state-space.R
\name{create_state_space}
\alias{create_state_space}
\title{Creates State Space}
\usage{
create_state_space(nrow = 3, ncol = 3)
}
\arguments{
\item{nrow}{height}

\item{ncol}{width}
}
\value{
an list of integers with class 4x5.space
}
\description{
Creates State Space
}
