% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_board.R
\name{evaluate_3x5_grid}
\alias{evaluate_3x5_grid}
\title{Title}
\usage{
evaluate_3x5_grid(G, LUT = xmas3_LUT_3x5, dims = c(3.5))
}
\arguments{
\item{G}{}

\item{LUT}{}

\item{dims}{}
}
\value{
data.frame with move suggestions
}
\description{
Title
}
