% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_board.R
\name{evaluate_board}
\alias{evaluate_board}
\title{Evaluate a board based in a Model}
\usage{
evaluate_board(B, q_model, dims = c(3, 5), include_strategy = TRUE)
}
\arguments{
\item{B}{The board}

\item{dims}{dimensions to use for windowing.}

\item{include_strategy}{should strategy be evaluated? (Slower)}

\item{model}{the model to use}
}
\value{
data.frame with suggested move and expected score
}
\description{
The function will create sequential windows on the board, and pass these through the model.
}
