% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_board.R
\name{evaluate_board_LUT}
\alias{evaluate_board_LUT}
\title{Evaluate Board Using Lookup Table}
\usage{
evaluate_board_LUT(B, LUT = Gatai_LUT, dims = c(3, 5))
}
\arguments{
\item{B}{Board to evaluate}

\item{LUT}{Lookup Table}

\item{dims}{dimension on model}
}
\value{
data.frame with suggested moves etc.
}
\description{
LUT is available and speeds up the evaluation process
}
