% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate-grid.R
\name{evaluate_grid}
\alias{evaluate_grid}
\title{Evauate a Sub-Grid}
\usage{
evaluate_grid(G)
}
\arguments{
\item{G}{The Grid to evaluate}
}
\description{
Calculates move sequences for a grid, a subset of a board.
}
