% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-all-sub-boards.R
\name{evaluate_sub_boards}
\alias{evaluate_sub_boards}
\title{Evaluate Sub-boards}
\usage{
evaluate_sub_boards(B, remove_duplicates = TRUE)
}
\arguments{
\item{B}{the list of sub-boards}

\item{remove_duplicates}{should duplciate move sequences be removed}
}
\value{
a data.frame with recommended moves
}
\description{
Evaluates a list of sub-boards
}
