% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highest-column-names.R
\name{find_column_with_highest_value}
\alias{find_column_with_highest_value}
\title{Return Column name With Highest Value}
\usage{
find_column_with_highest_value(data_frame)
}
\arguments{
\item{data_frame}{}
}
\value{
the column name
}
\description{
Return Column name With Highest Value
}
