% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatai-robot-class.R
\name{GataiRobot}
\alias{GataiRobot}
\title{Instance of Robot}
\description{
Handles the interaction with the web-driver.
}
\examples{
\dontrun{
## Play every legal move on the board.
gatai <- Robot$new()
gatai$clickToGame()
gatai$testRemoteDriver()
}

}
