% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-parallel-training-data.R
\name{generate_parallel_training_data}
\alias{generate_parallel_training_data}
\title{Generates Training Data for the Parallel State Space}
\usage{
generate_parallel_training_data()
}
\value{
a data.frame with State Action Reward NextState
}
\description{
Generates Training Data for the Parallel State Space
}
