% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-all-sub-boards.R
\name{get_all_sub_boards}
\alias{get_all_sub_boards}
\title{Get All Sub-boards}
\usage{
get_all_sub_boards(B, geometry = "3x3")
}
\arguments{
\item{B}{}

\item{geometry}{}
}
\value{
a list with the sub-boards
}
\description{
Calculates and returns all possible sub-boards for a given geometry.
}
