% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-and-save-individual-tiles.R
\name{get_and_save_individual_tiles}
\alias{get_and_save_individual_tiles}
\title{Get Individual tiles}
\usage{
get_and_save_individual_tiles(
  img,
  coordinates = null,
  output_path = here::here("inst", "image-data", "tiles"),
  filename = "image.png",
  size_x = 144,
  size_y = 144
)
}
\arguments{
\item{img}{the screenshot of the board}

\item{coordinates}{calculated centroids}

\item{output_path}{where to save the files}

\item{filename}{output filename}

\item{size_x}{dimensions}

\item{size_y}{dimensions}
}
\value{
invisible()
}
\description{
Chop the board up into individual tiles
}
