% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_combinations_and_score.R
\name{get_combinations_and_score}
\alias{get_combinations_and_score}
\title{Get Combinations}
\usage{
get_combinations_and_score(board)
}
\arguments{
\item{board}{}
}
\description{
Get Combinations
}
