% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-n-combinations.R
\name{get_n_combinations}
\alias{get_n_combinations}
\title{Number of Combinations}
\usage{
get_n_combinations(m)
}
\arguments{
\item{m}{a matrix representing the board.}
}
\value{
the sum of the row and column matches
}
\description{
Returns the number of combinations on the board, i.e. 3 or more consecutive
simlar tiles in a row or column.
}
