% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-get-windows.R
\name{get_windows}
\alias{get_windows}
\title{Get Windows for a Matrix}
\usage{
get_windows(M, direction = c("H", "V"), shape = c("1x4", "2x3"), size = 3)
}
\arguments{
\item{M}{the matrix on which to operate}

\item{direction}{}

\item{shape}{1 by4 or 2 by 3}

\item{size}{3 (for now)}
}
\value{
a data frame with w (window), col(start colum), row(start row)
}
\description{
Returns the windows for our matrix
Only works on 6x10
}
