% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize-board.R
\name{initialize_board}
\alias{initialize_board}
\title{Initialize a game board}
\usage{
initialize_board(avoid_combinations = TRUE)
}
\value{
a 6 by 10 Matrix with the \link{A-J} in the cells.
}
\description{
Initialize a game board
}
