% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_move.R
\name{make_move}
\alias{make_move}
\alias{make_move.xmas3}
\title{Make move on binary Board}
\usage{
make_move(x, ...)

\method{make_move}{xmas3}(B, move)
}
\arguments{
\item{B}{the board}

\item{move}{chess notation of the move i.e. "A1 to B1"}
}
\value{
the new board (a matrix)
}
\description{
Make move on binary Board
}
