% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{move}
\alias{move}
\title{Make a Move}
\usage{
move(board, .r = 1, .c = 1, move = c("U", "D", "L", "R"))
}
\arguments{
\item{board}{the current (or inital) board}

\item{.r}{row}

\item{.c}{column}

\item{move}{Up/Down/Left/Right}
}
\value{
The board after the move
}
\description{
Make a Move
}
