% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmas3-board-class.R
\name{parse_geometry}
\alias{parse_geometry}
\title{Parse Geometry}
\usage{
parse_geometry(g)
}
\arguments{
\item{g}{geometry to parse}
}
\value{
a two element numeric vector
}
\description{
Translate a geometry expressed as a character vector (e.g. "3x4") to a numeric representation.
(e.g. c(3,4))
}
