% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmas3-board-class.R
\name{parse_move}
\alias{parse_move}
\title{Parse Move}
\usage{
parse_move(move)
}
\arguments{
\item{move}{Character vector representing the move, e.g. "B1 to B2"}
}
\value{
a named list cell_specs c1,c2,r1,r2
}
\description{
Parse Move
}
