% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmas3-board-class.R
\name{parse_move_sequence}
\alias{parse_move_sequence}
\title{Parses a Move Sequence}
\usage{
parse_move_sequence(move_seq)
}
\arguments{
\item{move_seq}{}
}
\value{
a vector of the parsed moves
}
\description{
Parses a Move Sequence
}
