% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play-board.R
\name{play_board}
\alias{play_board}
\title{Make a Move on a Specific board}
\usage{
play_board(Board, agent = agent_0)
}
\arguments{
\item{Board}{a 6 by 10 Matrix with \link{A-J}}

\item{agent}{function which will make the moves. The default agent makes a random move on the board.}
}
\value{
a Matrix with the move made
}
\description{
Makes a move on a board.
}
