% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play-strategy.R
\name{play_strategy}
\alias{play_strategy}
\title{Play a Board with a Strategy}
\usage{
play_strategy(B, geometry = "3x3", strategy = c("bestmove"))
}
\arguments{
\item{B}{the board}

\item{geometry}{geometry of the strategy}

\item{strategy}{which one to choose (choose best move)}
}
\value{
move sequence(s) selected and expected score per move
}
\description{
Play a Board with a Strategy
}
