% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmas3-board-class.R
\name{plot.xmas3}
\alias{plot.xmas3}
\title{Plot Xmas3 Board}
\usage{
\method{plot}{xmas3}(x, ...)
}
\arguments{
\item{x}{Board to Plot}

\item{...}{}
}
\value{
a ggplot
}
\description{
Plot Xmas3 Board
}
