% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmas3-4x5-state.R
\name{print.4x5.space}
\alias{print.4x5.space}
\title{Print State 4x5 State Space}
\usage{
\method{print}{`4x5.space`}(x = 0, ...)
}
\arguments{
\item{x}{integer}

\item{...}{additional parameters to print function}
}
\value{
prints x as a binary matrix
}
\description{
Print State 4x5 State Space
}
