% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmas3-board-class.R
\name{print.xmas3}
\alias{print.xmas3}
\title{Print Xmas3 Board}
\usage{
\method{print}{xmas3}(x, ...)
}
\arguments{
\item{x}{board}

\item{...}{print parameters}
}
\value{
invisible NULL
}
\description{
Print Xmas3 Board
}
