% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robot-play.R
\name{robot_make_move}
\alias{robot_make_move}
\title{Make a Move}
\usage{
robot_make_move(.r = 1, .c = 1, move = c("U", "D", "L", "R"))
}
\arguments{
\item{.r}{row}

\item{.c}{Up Down Left Right}

\item{move}{move}

\item{offset_x}{canvas offset from board}

\item{offset_y}{canvas offset from board}
}
\value{
invisible()
}
\description{
Makes a move on the canvas
}
