% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robot-play.R
\name{robot_play_game}
\alias{robot_play_game}
\title{Play a Game through the UI}
\usage{
robot_play_game(agent = agent0, max_moves = 100)
}
\arguments{
\item{agent}{which agent to use}
}
\value{
invisible()
}
\description{
Play a Game through the UI
}
