% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score-and-update-board.R
\name{score_and_update_board}
\alias{score_and_update_board}
\title{Score and Update Board}
\usage{
score_and_update_board(board, with_score = TRUE)
}
\arguments{
\item{board}{}

\item{with_score}{}
}
\value{
list with updated board and score
}
\description{
Score and Update Board
}
