% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenshot-get-tile.R
\name{screenshot_get_tile}
\alias{screenshot_get_tile}
\title{Fecth tile from Screenshot}
\usage{
screenshot_get_tile(
  board,
  .c,
  .r,
  conf = robot_config_mac(),
  tile_width = 40,
  tile_height = 40
)
}
\arguments{
\item{board}{the board weŕe playing on}

\item{.c}{col}

\item{.r}{row}

\item{conf}{config}

\item{tile_width}{width}

\item{tile_height}{height}
}
\value{
the tile in png format
}
\description{
Fecth tile from Screenshot
}
