% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenshot-load-scale-and-crop.R
\name{screenshot_load_scale_and_crop}
\alias{screenshot_load_scale_and_crop}
\title{Scale and Crop Screenshot}
\usage{
screenshot_load_scale_and_crop(
  img,
  my_geometry = geometry_area(710, 426, 250, 134)
)
}
\arguments{
\item{img}{}

\item{my_geometry}{}
}
\value{
a magick image
}
\description{
Returns an image with only the board on.
}
