% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenshot-tile-guess-class.R
\name{screenshot_tile_guess_class}
\alias{screenshot_tile_guess_class}
\title{Guesses the Class of a Tile}
\usage{
screenshot_tile_guess_class(tile)
}
\arguments{
\item{tile}{The tile to analyze (magick image)}
}
\value{
class name of tile
}
\description{
Guesses the Class of a Tile
}
