% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmas3-board-class.R
\name{slide_move}
\alias{slide_move}
\title{Calculate Move on Different Board}
\usage{
slide_move(move, win)
}
\arguments{
\item{move}{the move}

\item{win}{window the move was played on.}
}
\value{
The Transitioned Move
}
\description{
When a move has been calculated on a sub-board and needs to be played on in a
different context (i.e. the full board) it needs to slide
}
