% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmas3-board-class.R
\name{slide_move_sequence}
\alias{slide_move_sequence}
\title{Slides a Sequence of Moves}
\usage{
slide_move_sequence(moves, pos)
}
\arguments{
\item{moves}{}

\item{pos}{}
}
\description{
Slides a Sequence of Moves
}
