% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmas3-board-class.R
\name{transpose}
\alias{transpose}
\alias{transpose.xmas3}
\title{Flip a Board}
\usage{
transpose(x, ...)

\method{transpose}{xmas3}(B)
}
\arguments{
\item{B}{xmas3board}
}
\value{
The flipped board
}
\description{
Flip a Board
}
