% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmas3-board-class.R
\name{transpose_move}
\alias{transpose_move}
\title{Transposes a Move}
\usage{
transpose_move(move)
}
\arguments{
\item{move}{Move to transpose}
}
\value{
the transposition
}
\description{
Makes vertical moves horizontal and vice-versa
}
