% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmas3-board-class.R
\name{window.xmas3}
\alias{window.xmas3}
\title{Returns a Window of the Board}
\usage{
\method{window}{xmas3}(B, from = c(1, 1), dims = c(2, 2))
}
\arguments{
\item{B}{Board to Window}

\item{dims}{window dimensions}

\item{pos}{Position}
}
\value{
a xmas3 board
}
\description{
Returns a Window of the Board
}
