% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmas3-board-class.R
\name{xmas3board}
\alias{xmas3board}
\title{Get Board State}
\usage{
xmas3board(x = 0, dims = c(4, 5))
}
\arguments{
\item{x}{integer}

\item{dim}{dimensions of the grid}
}
\value{
x with class and attributes added
}
\description{
Get Board State
}
