% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatai-robot-class.R
\name{GatAIRobot}
\alias{GatAIRobot}
\title{Instance of Robot}
\description{
Handles the interaction with the web-driver.
}
\examples{
\dontrun{
## Play every legal move on the board.
gatai <- Robot$new()
gatai$clickToGame()
gatai$testRemoteDriver()
}

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{takeScreenshot}}{Takes a screenshot of the current screen}

\item{\code{getBoardOnScreen}}{description
Takes a screenshot of the current screen, crops it and
returns the cropped board.}

\item{\code{play}}{Plays the current board using one of the GatAI agents.}

\item{\code{clickToGame}}{Clicks into the game from the welcome screen.}

\item{\code{getGameCanvas}}{Returns the binding to the game canvas.}

\item{\code{testRemoveDriver}}{Makes every legal move on the board. Convenience function for testing purposes.}

\item{\code{makeMove}}{Makes a move on the screen}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{takeScreenshot}}{Takes a screenshot of the current screen}

\item{\code{getBoardOnScreen}}{description
Takes a screenshot of the current screen, crops it and
returns the cropped board.}

\item{\code{play}}{Plays the current board using one of the GatAI agents.}

\item{\code{clickToGame}}{Clicks into the game from the welcome screen.}

\item{\code{getGameCanvas}}{Returns the binding to the game canvas.}

\item{\code{testRemoveDriver}}{Makes every legal move on the board. Convenience function for testing purposes.}

\item{\code{makeMove}}{Makes a move on the screen}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-x-mas-3-robot-new}{\code{GatAIRobot$new()}}
\item \href{#method-x-mas-3-robot-finalize}{\code{GatAIRobot$finalize()}}
\item \href{#method-x-mas-3-robot-message}{\code{GatAIRobot$message()}}
\item \href{#method-x-mas-3-robot-startRemoteDriver}{\code{GatAIRobot$startRemoteDriver()}}
\item \href{#method-x-mas-3-robot-takeScreenshot}{\code{GatAIRobot$takeScreenshot()}}
\item \href{#method-x-mas-3-robot-getBoardOnScreen}{\code{GatAIRobot$getBoardOnScreen()}}
\item \href{#method-x-mas-3-robot-play}{\code{GatAIRobot$play()}}
\item \href{#method-x-mas-3-robot-clickToGame}{\code{GatAIRobot$clickToGame()}}
\item \href{#method-x-mas-3-robot-getGameCanvas}{\code{GatAIRobot$getGameCanvas()}}
\item \href{#method-x-mas-3-robot-testRemoteDriver}{\code{GatAIRobot$testRemoteDriver()}}
\item \href{#method-x-mas-3-robot-makeMoveSequence}{\code{GatAIRobot$makeMoveSequence()}}
\item \href{#method-x-mas-3-robot-makeMove}{\code{GatAIRobot$makeMove()}}
\item \href{#method-x-mas-3-robot-clone}{\code{GatAIRobot$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-x-mas-3-robot-new"></a>}}
\if{latex}{\out{\hypertarget{method-x-mas-3-robot-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GatAIRobot$new(
  conf_filename = system.file("conf/chromebook.yml", package = "xmas3"),
  verbose = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-x-mas-3-robot-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-x-mas-3-robot-finalize}{}}}
\subsection{Method \code{finalize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GatAIRobot$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-x-mas-3-robot-message"></a>}}
\if{latex}{\out{\hypertarget{method-x-mas-3-robot-message}{}}}
\subsection{Method \code{message()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GatAIRobot$message(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-x-mas-3-robot-startRemoteDriver"></a>}}
\if{latex}{\out{\hypertarget{method-x-mas-3-robot-startRemoteDriver}{}}}
\subsection{Method \code{startRemoteDriver()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GatAIRobot$startRemoteDriver()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-x-mas-3-robot-takeScreenshot"></a>}}
\if{latex}{\out{\hypertarget{method-x-mas-3-robot-takeScreenshot}{}}}
\subsection{Method \code{takeScreenshot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GatAIRobot$takeScreenshot()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-x-mas-3-robot-getBoardOnScreen"></a>}}
\if{latex}{\out{\hypertarget{method-x-mas-3-robot-getBoardOnScreen}{}}}
\subsection{Method \code{getBoardOnScreen()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GatAIRobot$getBoardOnScreen()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-x-mas-3-robot-play"></a>}}
\if{latex}{\out{\hypertarget{method-x-mas-3-robot-play}{}}}
\subsection{Method \code{play()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GatAIRobot$play(agent = agent_0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{agent}}{a GatAi Agent function.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-x-mas-3-robot-clickToGame"></a>}}
\if{latex}{\out{\hypertarget{method-x-mas-3-robot-clickToGame}{}}}
\subsection{Method \code{clickToGame()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GatAIRobot$clickToGame(mode = c("arcade", "time"))}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-x-mas-3-robot-getGameCanvas"></a>}}
\if{latex}{\out{\hypertarget{method-x-mas-3-robot-getGameCanvas}{}}}
\subsection{Method \code{getGameCanvas()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GatAIRobot$getGameCanvas()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-x-mas-3-robot-testRemoteDriver"></a>}}
\if{latex}{\out{\hypertarget{method-x-mas-3-robot-testRemoteDriver}{}}}
\subsection{Method \code{testRemoteDriver()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GatAIRobot$testRemoteDriver()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-x-mas-3-robot-makeMoveSequence"></a>}}
\if{latex}{\out{\hypertarget{method-x-mas-3-robot-makeMoveSequence}{}}}
\subsection{Method \code{makeMoveSequence()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GatAIRobot$makeMoveSequence(moves)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-x-mas-3-robot-makeMove"></a>}}
\if{latex}{\out{\hypertarget{method-x-mas-3-robot-makeMove}{}}}
\subsection{Method \code{makeMove()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GatAIRobot$makeMove(move)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-x-mas-3-robot-clone"></a>}}
\if{latex}{\out{\hypertarget{method-x-mas-3-robot-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GatAIRobot$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
