% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate-grid.R
\name{get_sub_board}
\alias{get_sub_board}
\title{Get a Subset of a Board}
\usage{
get_sub_board(B, from, dims)
}
\arguments{
\item{B}{}

\item{from}{coordinates of the top-left cell}

\item{dims}{dimensions}
}
\value{
The sub-matrix
}
\description{
Get a Subset of a Board
}
