% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenshot-translate-board.R
\name{screenshot_translate_board}
\alias{screenshot_translate_board}
\title{Creates Matrix from Screenshot}
\usage{
screenshot_translate_board(B, conf = robot_config_mac())
}
\arguments{
\item{B}{the screenshot}
}
\value{
a 6x10 Matrix
}
\description{
Creates Matrix from Screenshot
}
